Factor out c_analyzer_common package.
