+++ Lib/test/test_tools/test_c_analyzer/util.py
